fix: fix invalid dist in bonus

+++ cub3D_bonus/src/raycast/helper.c
@@ -6,7 +6,7 @@
 /*   By: mlitvino <mlitvino@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/25 18:59:38 by mlitvino          #+#    #+#             */
-/*   Updated: 2025/06/26 18:57:27 by mlitvino         ###   ########.fr       */
+/*   Updated: 2025/06/26 19:13:30 by mlitvino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	calc_norm_dist(t_raycast *raycast)
 {
-	double	temp;
+	long double	temp;
 
 	if (raycast->hor_wall.dist != INT_MAX)
 	{
